final/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.22
sentence-transformers==2.2.2
torch>=2.1.0
PyMuPDF==1.23.8
Pillow==10.0.1
python-multipart==0.0.6
pyyaml>=6.0.1
tqdm>=4.66.1
aiofiles>=23.2.1
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
